split files into two files. all helper methods are moved into different file.
