+++ src/Compilers/CSharp/Portable/CSharpCodeAnalysis.csproj
@@ -451,6 +451,7 @@
     <Compile Include="Lowering\SyntheticBoundNodeFactory.cs" />
     <Compile Include="Lowering\UnmatchedGotoFinder.cs" />
     <Compile Include="Operations\CSharpOperationFactory.cs" />
+    <Compile Include="Operations\CSharpOperationFactory_Methods.cs" />
     <Compile Include="Parser\AbstractLexer.cs" />
     <Compile Include="Parser\BlendedNode.cs" />
     <Compile Include="Parser\Blender.cs" />
